move reaper thread to DCP extension
