+++ extensions/common/store/sql/jti-validation-store-sql/src/test/java/org/eclipse/edc/edr/store/index/sql/SqlJtiValidationStoreExtensionTest.java
@@ -55,14 +55,4 @@ public class SqlJtiValidationStoreExtensionTest {
 
         verify(config).getString(eq(DATASOURCE_NAME), any());
     }
-
-    @Test
-    void shouldStartReaperThread_withDefaultConfig(SqlJtiValidationStoreExtension extension, ServiceExtensionContext context) {
-        var config = mock(Config.class);
-        when(context.getConfig()).thenReturn(config);
-        when(config.getLong(SqlJtiValidationStoreExtension.CLEANUP_PERIOD, any())).thenReturn(1L);
-
-        extension.initialize(context);
-        extension.start();
-    }
 }
